refactor: auth 모듈로 credentials 통합
