+++ server/src/main/java/com/emr/slgi/auth/dto/CredentialsCreateDTO.java
@@ -1,4 +1,4 @@
-package com.emr.slgi.credentials.dto;
+package com.emr.slgi.auth.dto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
